requirements.txt
py-cord
python-dotenv
sqlalchemy
asyncpg